Split test/lib/questions.test.ts, drop its biome override
Removed test/lib/questions.test.ts (1195 lines) from biome.json's
noExcessiveLinesPerFile override list. Split it into a test/lib/questions/
directory: crud.test.ts, listing-mapping.test.ts, with-listings.test.ts,
attendee-answers.test.ts, and a shared helpers.ts (createAttendee,
createQuestion, addAnswer, createQuestionWithAnswers, saveTextAnswers,
expectQuestionTexts, createOrderedQuestionPair,
seedQuestionWithAndWithoutAnswers).

Splitting the oversized file let jscpd scan it for the first time,
surfacing extensive duplication in the question/answer setup boilerplate
that had been invisible inside the 1195-line monolith (20 clones on first
pass). The repeated "insert a radio question, insert its answers" and
"save an attendee's free-text answers" shapes are now factored into the
helpers above; the handful of tests that shared identical multi-line setup
beyond that (two questions across two listings, a question assigned to two
listings, a with/without-answers listing fixture, a globally-ordered
question pair) got their own local seed helpers. Ends at 0 clones.
